Fix issue with installation of gfortran 11
In my repo I had issues running a program compiled with gfortran 11 in
CI. The problem turned out to be that it was successfully compiled with
gfortran 11, but it was linking against the wrong libgfortran-5.dll at
runtime. It was linking to the one in /c/mingw64/bin, which is from
gfortran 12 which is installed by default in the Windows 2022 image.

The simple solution is to move the entire mingw64 folder instead of just
a few files.