refactor(gcc): drop support for gcc 7 and 8 (#83)
These have patchy availability. And anyways, the oldest release series still receiving support is 11, so there may be an argument for dropping 9 and 10, but I will leave that for later.

In any case, nothing prevents trying unsupported versions, but we will no longer test them or make guarantees. This frees up room in the test matrix, which is capped at 256 jobs