Sync fork (#19)
* fix(mac/lfortran): only install micromamba if needed (#79)

* docs(readme): mention lfortran in compiler option section (#80)

* ci: include long_compat.csv in autoupdate prs (#85)

* Update compatibility matrix (#86)

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>

* refactor(gcc): drop support for gcc 7 and 8 (#83)

These have patchy availability. And anyways, the oldest release series still receiving support is 11, so there may be an argument for dropping 9 and 10, but I will leave that for later.

In any case, nothing prevents trying unsupported versions, but we will no longer test them or make guarantees. This frees up room in the test matrix, which is capped at 256 jobs

* feat(intel): support intel 2024.1 on linux and windows (#84)

---------

Co-authored-by: wpbonelli <wbonelli@ucar.edu>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>